+++ packages/gatsby-source-wordpress/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wordpress@3.3.2...gatsby-source-wordpress@3.3.3) (2020-05-13)
+
+**Note:** Version bump only for package gatsby-source-wordpress
+
 ## [3.3.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-wordpress@3.3.1...gatsby-source-wordpress@3.3.2) (2020-05-13)
 
 **Note:** Version bump only for package gatsby-source-wordpress